experiments/two-ways-of-repeated-inference/requirements.txt
opencc-python-reimplemented
psutil
requests

